Game flow (#13)
* Game flow

Added DOTween and UniTask to the project, created a LevelManager for the scene transitions which consists of a simple fading effect using DOTween and Unitask. Added a trigger condition when the ball goes into the hole. Also created a splash screen.

* Modified it so it won't be async

Used LoadSceneAsync(sceneName).Forget() just as told

* Removed unused UniTask library from the script

* Script name change and .onClick addition

Changed the script name to ChangeSceneButton and implemented .OnClick directly into the code just as suggested.

* Removed OnClick references from each button

Removed each OnClick references in each scene so it uses the event directly from code

* Minor adjustments

---------

Co-authored-by: Ivan Herrera <ivan.herrera.m@outlook.com>