Added Fonts, UI and Prefabs
UI assets were added, fonts needed for buttons, texts and titles.
UI prefabs that we will need were created, this includes buttons and panels. I also have to be honest, I think there are things left to do but I didn't know what else and I had to deliver them so as not to delay the team.