Save System created
The save system was created. I made 3 scripts, 2 are about the operation of the system and the third is an example. The first script "GameData" is the basis of the information that we want to save about the Player or the game, the second script "SaveSystem" is to save the player information in a Json document and the third script is an example of how to call the stored information and how to save it again.