Tutorial mechanics (#33)
* Tutorial art addition

Added missing art

* Tilemaps creation

* InputManager and Player prefabs creation

* Tutorial and Testing scene modifications

Added tutorial scene to the build settings and changed its art. Modified some things from the testing scene.

* Level selector script addition

Made a level selection script so it unlocks each level depending on the game progression. It only works with PlayerPrefs for now, can be changed later on to work with the json Save system.

* End Level Trigger script PlayerPrefs

Modified the end level trigger script so it sets the specific value to player prefs and the next level can be unlocked.

* Small script change

Changed the rigidbody GetComponent to Awawke instead of Start

* Small tutorial scene changes

* Minor changes