Gif player was added (#24)
* Gif player was added

The plugin to use GIFs was added, I created a scene to test it and also make a small example, it is a GameObject to which we add Sprite Renderer and a script that comes in the plugin, Gif Player which will allow us to bring the gif to life when we run the program.

* Save System created

The save system was created. I made 3 scripts, 2 are about the operation of the system and the third is an example. The first script "GameData" is the basis of the information that we want to save about the Player or the game, the second script "SaveSystem" is to save the player information in a Json document and the third script is an example of how to call the stored information and how to save it again.

* Fixed issue with SaveSystem

Name of the class "EXAMPLE" changed, is now named PlayerData, in this script we can save the Data that we need of the game or the user.
The missing space lines were added.
Comments were deleted, no more spanish.
Singleton Implemented and the interface too.
The interface have one method for each variable we want to save or load.

Also here is the evidence in video of the GIF player: https://drive.google.com/file/d/17kkzUvJ3JxJow1zFrhoFs_D_zUq7x71q/view?usp=drive_link

* Save System fixed

Script PlayerData was deleted due to the scrip was merged to SaveSystem script.
Methods writted in multiple lines, also I deleted the variables that we are not using like the player position and the score of the player.

* Save System Fixed

The changes required for the save system were implemented, and the errors that caused information to be deleted were eliminated.

* Removed uselless method

---------

Co-authored-by: Ivan Herrera <ivan.herrera.m@outlook.com>