Fixes and improvements (#175)
* New input events added

* MenuTransitionController script

Script in charge of all of the transitions of the main menu, all of them are done with DOTween and UniTask

* New SlotButton script

Same as SlotsButton but had to test this new menu

* LoadGameButton script

* ScrollingTextController + MenuButtonSelect scripts

MenuButtonSelect allows the user to set up the text to be scrolled depending on the selected button

* ConfirmationPanel modification

Adapted it so it works with controller

* Scene modifications

* MenuTransitionController small script changes

* SlotButton script fix

* Moved some lines into a sequence

* Font assets addition

* MenuTransitionController settings adaptation

* Main menu buttons scripts

* ConfirmationPanel fix

* Localization fixes

* SaveSystem localization

* Scene update

* Font addition

* Background addition

* ExitButton script creation

* Scene modification

* Scenes update

* Added missing scenes to build settings

* Updated files

* Changed UI manager prefab

* Added controller inputs to InputManager

* Scripts update

* Chaneges

* AudioSettings fix

* Credits music

* Credits transition

* save hits

* save scene

* Fixed dotween error coming from main menu transition

* Fixed input manager persistent inconsistency

* LevelSelector CRITICAL BUG FIX

Fixed bug which caused a soft reset when the game was paused on the LevelSelector

* Localization bug fix

Fixed bug where the Localization.csv file was missing in the build. Moved the file to resources and changed a few lines in the LocalizationManager script so it can read the file.

* CRITICAL BUG FIX

Critical bug which caused a soft reset during the LevelSelector and Tutorial because of the InputManager and GameStateManager. Fixed bug of pause during dialogues. Added a new game state and set the corresponding conditions.

* Main menu music addition

Added main menu music asset and database entry

* AUDIO BUG FIX

Fixed bug where the audio wouldn't load properly even though the audio was saved it wouldn't play on certain scenes and when the game was reopened the sliders would set back to 0 even though there was a saved settings value

* Prefabs adjustments

* UI scripts adjustments

* Scenes adjustments

* Saved current strokes number

* Small LevelManager fix

* Fonts assets

---------

Co-authored-by: Ivan Herrera <ivan.herrera.m@outlook.com>